Removed unneeded options on main menu and gameplay menu
